Extracted commonalities from Phone, Email and Address classes into a parent class called Contact.
